ssc setup/fte: Don't create .ssc.env if vcredist not install
Ensures that FTE prompt runs each time ssc.js is called
Without this, ssc.exe fails silently
(vcruntime140_1.dll required)

+++ npm/src/index.js
@@ -51,7 +51,7 @@ export const firstTimeExperienceSetup = async () => {
   })
 
   // If fte didn't create a configuration file, make an empty one to prevent user being prompted again
-  if (!fs.existsSync(path.join(installPath, '.ssc.env'))) {
+  if (child.exitCode == 0 && !fs.existsSync(path.join(installPath, '.ssc.env'))) {
     fs.writeFileSync(path.join(installPath, '.ssc.env'), "")
   }
 }
